clGen: Subparm default handling: fall back to \rest rather than nil
If there was no default and no events, previously it would generate
an event with no 'item' -- causing the phrase pattern to die on the
first event with nil and stop the whole process. Now it substitutes
\rest in that case, which causes the whole phrase to go silent
but this is at least not catastrophic (remedy it and the process
will start playing again at the next barline). Post a warning.