pitch-procs: Fix chordBP to prefer user-supplied mode
Desired behavior in LC improv is to adapt pitch material, originally
given in one mode, to a new mode on the fly. Chucklib objects
preferred to keep the integrity of the given MIDIRecBuf. Here,
we use a new override to analyze the notes by the MIDIRecBuf's own
mode, but then play it back using the process's mode.

arpegBP is not fixed yet.