Pattern wrappers now use the handler directly for conversion
Previously, pattern wrappers called back into the BP to convert values
because the BP had access to the parmMap. Now the parm handler object
has access to this. So there's no need to call back up to the parent.
This means it's no longer necessary to pass the original parameter key
for value conversion either, so 'inParm' is now gone.