Remove extraneous 'parm' name/array passing
The parm handler knows everything about the parm, and this shouldn't change.
So we don't have to figure out, for every pattern-set, whether it's
storing as 'parm' or '[parm, \delta, \dur]' and we don't have to look up
aliases every time. Let the handler do it.

So where it was passing parm *and* storeParm *and* inParm, now we don't
need to pass *any* parm names. 'handler.wrapPattern' already knows everything.